make Encoding it's own class
