+++ backend/src/main/java/com/bakdata/conquery/commands/CollectEntitiesCommand.java
@@ -6,7 +6,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -19,10 +18,9 @@ import java.util.zip.GZIPInputStream;
 import com.bakdata.conquery.ConqueryConstants;
 import com.bakdata.conquery.models.datasets.Dataset;
 import com.bakdata.conquery.models.dictionary.EncodedDictionary;
-import com.bakdata.conquery.models.events.stores.specific.string.StringTypeEncoded;
+import com.bakdata.conquery.models.dictionary.Encoding;
 import com.bakdata.conquery.models.exceptions.JSONException;
 import com.bakdata.conquery.models.jobs.SimpleJob.Executable;
-import com.bakdata.conquery.models.preproc.Preprocessed;
 import com.bakdata.conquery.models.preproc.PreprocessedDictionaries;
 import com.bakdata.conquery.models.preproc.PreprocessedHeader;
 import com.bakdata.conquery.models.preproc.PreprocessedReader;
@@ -125,7 +123,7 @@ public class CollectEntitiesCommand extends Command {
 
 				final PreprocessedDictionaries dictionaries = parser.readDictionaries();
 
-				final EncodedDictionary primaryDictionary = new EncodedDictionary(dictionaries.getPrimaryDictionary(), StringTypeEncoded.Encoding.UTF8);
+				final EncodedDictionary primaryDictionary = new EncodedDictionary(dictionaries.getPrimaryDictionary(), Encoding.UTF8);
 
 				add(primaryDictionary, new File(file.getParentFile(), "all_entities.csv"));
 				if (verbose) {
