+++ backend/src/main/java/com/bakdata/conquery/models/dictionary/EncodedDictionary.java
@@ -1,7 +1,5 @@
 package com.bakdata.conquery.models.dictionary;
 
-import com.bakdata.conquery.models.events.stores.specific.string.StringTypeEncoded;
-
 /**
  * Handle class allowing direct encoded access to Dictionary without wrapping inside StringTypeEncoded.
  *
@@ -10,9 +8,9 @@ import com.bakdata.conquery.models.events.stores.specific.string.StringTypeEncod
 public class EncodedDictionary {
 
 	private final Dictionary dict;
-	private final StringTypeEncoded.Encoding encoding;
+	private final Encoding encoding;
 
-	public EncodedDictionary(Dictionary dict, StringTypeEncoded.Encoding encoding) {
+	public EncodedDictionary(Dictionary dict, Encoding encoding) {
 		this.dict = dict;
 		this.encoding = encoding;
 	}
