+++ backend/src/main/java/com/bakdata/conquery/models/events/stores/specific/string/StringTypeEncoded.java
@@ -1,6 +1,5 @@
 package com.bakdata.conquery.models.events.stores.specific.string;
 
-import java.nio.charset.StandardCharsets;
 import java.util.Iterator;
 
 import javax.annotation.Nonnull;
@@ -8,6 +7,7 @@ import javax.annotation.ParametersAreNonnullByDefault;
 
 import com.bakdata.conquery.io.cps.CPSType;
 import com.bakdata.conquery.models.dictionary.Dictionary;
+import com.bakdata.conquery.models.dictionary.Encoding;
 import com.bakdata.conquery.models.events.stores.root.ColumnStore;
 import com.bakdata.conquery.models.events.stores.root.IntegerStore;
 import com.bakdata.conquery.models.events.stores.root.StringStore;
@@ -16,8 +16,10 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
-import com.google.common.io.BaseEncoding;
-import lombok.*;
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.Setter;
+import lombok.SneakyThrows;
 
 /**
  * Compacted String store, that uses two methods to reduce memory footprint:
@@ -165,59 +167,4 @@ public class StringTypeEncoded implements StringStore {
 		return subType.has(event);
 	}
 
-	/**
-	 * We use common Encodings in the reversed way. What the encoding sees as "encoded" data,
-	 * is actually our raw data. On this raw data the decoding of the chosen encoding applied, which
-	 * yield a smaller representation for storage in the memory.
-	 *
-	 * To use this technique all string in the dictionary must only use the dictionary that is inherent
-	 * to the chosen encoding.
-	 */
-	@RequiredArgsConstructor
-	public static enum Encoding {
-		// Order is for precedence, least specific encodings go last.
-		Base16LowerCase(2, BaseEncoding.base16().lowerCase().omitPadding()),
-		Base16UpperCase(2, BaseEncoding.base16().upperCase().omitPadding()),
-		Base32LowerCase(8, BaseEncoding.base32().lowerCase().omitPadding()),
-		Base32UpperCase(8, BaseEncoding.base32().upperCase().omitPadding()),
-		Base32HexLowerCase(8, BaseEncoding.base32Hex().lowerCase().omitPadding()),
-		Base32HexUpperCase(8, BaseEncoding.base32Hex().upperCase().omitPadding()),
-		Base64(4, BaseEncoding.base64().omitPadding()),
-		UTF8(1, null) {
-			@Override
-			public String decode(byte[] bytes) {
-				return new String(bytes, StandardCharsets.UTF_8);
-			}
-
-			@Override
-			public byte[] encode(String chars) {
-				return chars.getBytes(StandardCharsets.UTF_8);
-			}
-
-			@Override
-			public boolean canEncode(String chars) {
-				return true;
-			}
-		};
-
-		private final int requiredLengthBase;
-		private final BaseEncoding encoding;
-
-		public String decode(byte[] bytes) {
-			// Using encode here is valid see comment on this enum
-			return encoding.encode(bytes);
-		}
-
-		public boolean canEncode(String chars) {
-			// Using canDecode here is valid see comment on this enum
-			return encoding.canDecode(chars)
-				   && chars.length() % requiredLengthBase == 0;
-		}
-
-		public byte[] encode(String chars) {
-			// Using decode here is valid see comment on this enum
-			return encoding.decode(chars);
-		}
-
-	}
 }
