+++ backend/src/main/java/com/bakdata/conquery/models/preproc/Preprocessed.java
@@ -14,10 +14,10 @@ import java.util.zip.GZIPOutputStream;
 import com.bakdata.conquery.io.jackson.Jackson;
 import com.bakdata.conquery.models.config.ConqueryConfig;
 import com.bakdata.conquery.models.dictionary.Dictionary;
+import com.bakdata.conquery.models.dictionary.Encoding;
 import com.bakdata.conquery.models.events.MajorTypeId;
 import com.bakdata.conquery.models.events.stores.root.ColumnStore;
 import com.bakdata.conquery.models.events.stores.root.StringStore;
-import com.bakdata.conquery.models.events.stores.specific.string.StringTypeEncoded;
 import com.bakdata.conquery.models.preproc.parser.ColumnValues;
 import com.bakdata.conquery.models.preproc.parser.Parser;
 import com.bakdata.conquery.models.preproc.parser.specific.StringParser;
@@ -216,7 +216,7 @@ public class Preprocessed {
 	private Dictionary encodePrimaryDictionary() {
 		log.debug("Encode primary Dictionary");
 
-		primaryColumn.applyEncoding(StringTypeEncoded.Encoding.UTF8);
+		primaryColumn.applyEncoding(Encoding.UTF8);
 
 		final Dictionary primaryDictionary = new MapTypeGuesser(primaryColumn).createGuess().getType().getUnderlyingDictionary();
 		log.trace("\tPrimaryColumn -> {}", primaryDictionary);
