+++ backend/src/test/java/com/bakdata/conquery/io/storage/xodus/stores/BigStoreTest.java
@@ -14,8 +14,8 @@ import com.bakdata.conquery.models.config.XodusStoreFactory;
 import com.bakdata.conquery.models.datasets.Dataset;
 import com.bakdata.conquery.models.dictionary.Dictionary;
 import com.bakdata.conquery.models.dictionary.EncodedDictionary;
+import com.bakdata.conquery.models.dictionary.Encoding;
 import com.bakdata.conquery.models.dictionary.MapDictionary;
-import com.bakdata.conquery.models.events.stores.specific.string.StringTypeEncoded;
 import com.bakdata.conquery.models.exceptions.JSONException;
 import com.bakdata.conquery.models.identifiable.CentralRegistry;
 import com.bakdata.conquery.models.identifiable.ids.specific.DictionaryId;
@@ -95,7 +95,7 @@ public class BigStoreTest {
 				store.getMetaStore().get(nDict.getId()).loadData(store.getDataStore()).map(ByteArrayInputStream::new).iterator())))
 					.hasSameContentAs(new ByteArrayInputStream(bytes));
 
-		EncodedDictionary copy = new EncodedDictionary(store.get(nDict.getId()), StringTypeEncoded.Encoding.UTF8);
+		EncodedDictionary copy = new EncodedDictionary(store.get(nDict.getId()), Encoding.UTF8);
 		for (int v = 0; v < 1000000; v++) {
 			assertThat(copy.getId(Integer.toHexString(v))).isEqualTo(v);
 		}
